% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calendars.R
\name{calendar_td}
\alias{calendar_td}
\title{Calendar Specific Trading Days Variables}
\usage{
calendar_td(
  calendar,
  frequency,
  start,
  length,
  s,
  groups = c(1, 2, 3, 4, 5, 6, 0),
  holiday = 7,
  contrasts = TRUE,
  meanCorrection = contrasts
)
}
\arguments{
\item{calendar}{The calendar.}

\item{frequency}{Annual frequency (divisor of 12).}

\item{start, length}{First date (array with the first year and the first period)
(for instance `c(1980, 1)`) and number of periods of the output variables. Can also be provided with the `s` argument}

\item{s}{time series used to get the dates for the trading days variables. If supplied the
parameters `frequency`, `start` and `length` are ignored.}

\item{groups}{Groups of days. The length of the array must be 7. It indicates to what group each week day
belongs. The first item corresponds to Mondays and the last one to Sundays. The group used for contrasts (usually Sundays) is identified by 0.
The other groups are identified by 1, 2,... n (<= 6). For instance, usual trading days are defined by c(1,2,3,4,5,6,0),
week days by c(1,1,1,1,1,0,0), week days, Saturdays, Sundays by c(1,1,1,1,1,2,0) etc...}

\item{holiday}{Day for holidays (holidays are considered as that day).
1 for Monday... 7 for Sunday. Doesn't necessary belong to the 0-group.}

\item{contrasts}{If true, the variables are defined by contrasts with the 0-group. Otherwise, raw number of days are provided}

\item{meanCorrection}{boolean indicating if the regressors are corrected for long-term term.
By default the correction is done if \code{contrasts = TRUE}.}
}
\value{
The variables corresponding to each group, starting with the 0-group (\code{contrasts = FALSE})
or the 1-group (\code{contrasts = TRUE}).
}
\description{
Calendar Specific Trading Days Variables
}
\examples{
BE <- national_calendar(list(
    fixed_day(7,21),
    special_day('NEWYEAR'),
    special_day('CHRISTMAS'),
    special_day('MAYDAY'),
    special_day('EASTERMONDAY'),
    special_day('ASCENSION'),
    special_day('WHITMONDAY'),
    special_day('ASSUMPTION'),
    special_day('ALLSAINTSDAY'),
    special_day('ARMISTICE')))
calendar_td(BE, 12, c(1980,1), 240)
}
