% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calendars.R
\name{chained_calendar}
\alias{chained_calendar}
\title{Chained Calendar}
\usage{
chained_calendar(calendar1, calendar2, break_date)
}
\arguments{
\item{calendar1, calendar2}{calendars to chain.}

\item{break_date}{the break date in the format \code{"YYYY-MM-DD"}.}
}
\value{
The chained calendar.
}
\description{
A chained calendar is an option that can be used when major changes in the composition of the holidays take place.
In such case two calendars that describe the situation before and after the change of regime can be defined.
}
