% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{densityT}
\alias{densityT}
\title{Title}
\usage{
densityT(df, x)
}
\arguments{
\item{x}{}
}
\value{

}
\description{
Title
}
\examples{
# T with 2 degrees of freedom.
z<-densityT(2, .01*seq(-100, 100, 1))
}
