% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/procresults.R
\name{dictionary}
\alias{dictionary}
\alias{result}
\title{Get Dictionary and Result}
\usage{
dictionary(object)

result(object, id)
}
\arguments{
\item{object}{the object.}

\item{id}{the name of the object to extract.}
}
\description{
Extract dictionnay of a \code{"JD3_ProcResults"} object (\code{dictionary()}) and extract a specific value (\code{result()})
}
