% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calendars.R
\name{easter_dates}
\alias{easter_dates}
\title{Compute Easter days between two years}
\usage{
easter_dates(year0, year1, julian = FALSE)
}
\arguments{
\item{year0, year1}{years.}

\item{julian}{boolean indicating if julian calendar must be used.}
}
\description{
Compute Easter days between two years
}
\examples{
easter_dates(2020, 2021)
}
\seealso{
Other calendar functions: 
\code{\link{calendar_td}()},
\code{\link{easter_day}()},
\code{\link{fixed_day}()},
\code{\link{fixed_week_day}()},
\code{\link{holidays}()},
\code{\link{long_term_mean}()},
\code{\link{national_calendar}()},
\code{\link{single_day}()},
\code{\link{special_day}()},
\code{\link{stock_td}()},
\code{\link{td}()}
}
\concept{calendar functions}
