% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variables.R
\name{easter_variable}
\alias{easter_variable}
\alias{julianeaster_variable}
\title{Easter regressors}
\usage{
easter_variable(
  frequency,
  start,
  length,
  s,
  duration = 6,
  endpos = -1,
  correction = c("Simple", "PreComputed", "Theoretical", "None")
)

julianeaster_variable(frequency, start, length, s, duration = 6)
}
\arguments{
\item{frequency}{Annual frequency (divisor of 12).}

\item{start, length}{First date (array with the first year and the first period)
(for instance `c(1980, 1)`) and number of periods of the output variables. Can also be provided with the `s` argument}

\item{s}{time series used to get the dates for the trading days variables. If supplied the
parameters `frequency`, `start` and `length` are ignored.}

\item{duration}{Duration (length in days) of the Easter effect.}

\item{endpos}{Position of the end of the Easter effect, relatively to Easter.}

\item{correction}{mean correction option.}
}
\description{
Easter regressors
}
