% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calendars.R
\name{fixed_day}
\alias{fixed_day}
\title{Fixed Day}
\usage{
fixed_day(month, day, weight = 1, validity = NULL)
}
\arguments{
\item{month, day}{the month and the day of the fixed day to add.}

\item{weight}{weight associated to the holiday.}

\item{validity}{validity period: either \code{NULL} (full sample) or a named list with \code{"start"} and/or "end" dates in the format \code{"YYYY-MM-DD"}.}
}
\description{
Fixed Day
}
\examples{
day <- fixed_day(7, 21, .9)
day  # 21 July, with weight=0.9
day <- fixed_day(12, 25, .5, validity = list(start = "2010-01-01"))
day # 21 July, with weight=0.5 from Jan 2010
day <- fixed_day(12, 25, .5, validity = list(end = "2010-01-01"))
day # 21 July, with weight=0.9 until Jan 2010
}
\seealso{
Other calendar functions: 
\code{\link{calendar_td}()},
\code{\link{easter_dates}()},
\code{\link{easter_day}()},
\code{\link{fixed_week_day}()},
\code{\link{holidays}()},
\code{\link{long_term_mean}()},
\code{\link{national_calendar}()},
\code{\link{single_day}()},
\code{\link{special_day}()},
\code{\link{stock_td}()},
\code{\link{td}()}
}
\concept{calendar functions}
