% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calendars.R
\name{fixed_week_day}
\alias{fixed_week_day}
\title{Fixed Week Day}
\usage{
fixed_week_day(month, week, dayofweek, weight = 1, validity = NULL)
}
\arguments{
\item{month}{month of the holiday: from `1` (January) to `12` (December).}

\item{week}{position of the specified week day in the month: from `1` (first week of the month) to `5`. Should be always lower than 5.
`-1` for the last `dayofweek` of the month.}

\item{dayofweek}{day of the week: from `1` (Monday) to `7` (Sunday).}

\item{weight}{weight associated to the holiday.}

\item{validity}{validity period: either `NULL` (full sample) or a named list with `"start"` and/or "end" dates in the format `"YYYY-MM-DD"`.}
}
\description{
Fixed Week Day
}
\examples{
day <- fixed_week_day(7, 2, 3) # second Tuesday of July
day

}
