% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{invgammadistribution}
\alias{invgammadistribution}
\alias{randomsInverseGamma}
\alias{densityInverseGamma}
\alias{cdfInverseGamma}
\title{The Inverse-Gamma Distribution}
\usage{
randomsInverseGamma(shape, scale, n)

densityInverseGamma(shape, scale, x)

cdfInverseGamma(shape, scale, x)
}
\arguments{
\item{shape}{shape and scale parameters.}

\item{scale}{shape and scale parameters.}

\item{n}{number of observations.}

\item{x}{vector of quantiles.}
}
\description{
Density, (cumulative) distribution function and random generation for inverse-gamma distribution.
}
