% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{invgaussiandistribution}
\alias{invgaussiandistribution}
\alias{randomsInverseGaussian}
\alias{densityInverseGaussian}
\alias{cdfInverseGaussian}
\title{The Inverse-Gaussian Distribution}
\usage{
randomsInverseGaussian(shape, scale, n)

densityInverseGaussian(shape, scale, x)

cdfInverseGaussian(shape, scale, x)
}
\arguments{
\item{shape}{shape and scale parameters.}

\item{scale}{shape and scale parameters.}

\item{n}{number of observations.}

\item{x}{vector of quantiles.}
}
\description{
Density, (cumulative) distribution function and random generation for inverse-gaussian distribution.
}
