% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood.R
\name{likelihood}
\alias{likelihood}
\title{Generic Likelihood Function for 'JDemetra+'}
\usage{
likelihood(
  nobs,
  neffectiveobs = NA,
  nparams = 0,
  ll,
  adjustedll = NA,
  aic,
  aicc,
  bic,
  bicc,
  ssq
)
}
\arguments{
\item{nobs, neffectiveobs, nparams, ll, adjustedll, aic, aicc, bic, bicc, ssq}{parameters.}
}
\value{
A \code{"JD3_LIKELIHOOD"} object.
}
\description{
Generic functions to format likelihood 'JDemetra+'.
}
