% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calendars.R
\name{national_calendar}
\alias{national_calendar}
\title{Create National Calendar}
\usage{
national_calendar(days)
}
\arguments{
\item{days}{list of days of the calendars}
}
\description{
Create National Calendar
}
\examples{
BE <- national_calendar(list(
    fixed_day(7,21),
    special_day('NEWYEAR'),
    special_day('CHRISTMAS'),
    special_day('MAYDAY'),
    special_day('EASTERMONDAY'),
    special_day('ASCENSION'),
    special_day('WHITMONDAY'),
    special_day('ASSUMPTION'),
    special_day('ALLSAINTSDAY'),
    special_day('ARMISTICE')))
}
\seealso{
Other calendar functions: 
\code{\link{calendar_td}()},
\code{\link{easter_dates}()},
\code{\link{easter_day}()},
\code{\link{fixed_day}()},
\code{\link{fixed_week_day}()},
\code{\link{holidays}()},
\code{\link{long_term_mean}()},
\code{\link{single_day}()},
\code{\link{special_day}()},
\code{\link{stock_td}()},
\code{\link{td}()}
}
\concept{calendar functions}
