% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{randomsT}
\alias{randomsT}
\title{Title}
\usage{
randomsT(df, n)
}
\arguments{
\item{n}{}
}
\value{

}
\description{
Title
}
\examples{
# T with 2 degrees of freedom. 100 randoms
z<-randomsT(2, 100)
}
