% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arima.R
\name{sarima.decompose}
\alias{sarima.decompose}
\title{Title}
\usage{
sarima.decompose(model, rmod = 0, epsphi = 0)
}
\arguments{
\item{model}{Sarima model to decompose}

\item{rmod}{Trend threshold}

\item{epsphi}{Seasonal tolerance (in degrees)}
}
\description{
Title
}
\examples{
model <- sarima.model(period = 12, d =1, bd = 1, theta = -0.6, btheta = -0.5)
ucm <- sarima.decompose(model)

}
