% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arima.R
\name{sarima.estimate}
\alias{sarima.estimate}
\title{Title}
\usage{
sarima.estimate(
  data,
  order = c(0, 0, 0),
  seasonal = list(order = c(0, 0, 0), period = 1),
  mean = FALSE,
  xreg = NULL,
  eps = 1e-09
)
}
\arguments{
\item{eps}{}
}
\description{
Title
}
