% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arima.R
\name{sarima.model}
\alias{sarima.model}
\title{Seasonal ARIMA model (Box-Jenkins)}
\usage{
sarima.model(
  name = "sarima",
  period,
  phi = NULL,
  d = 0,
  theta = NULL,
  bphi = NULL,
  bd = 0,
  btheta = NULL
)
}
\arguments{
\item{name}{name of the model.}

\item{period}{period of the model.}

\item{phi}{coefficients of the regular auto-regressive polynomial (1 + phi(1)B + phi(2)B + ...). True signs.}

\item{d}{regular differencing order.}

\item{theta}{coefficients of the regular moving average polynomial (1 + theta(1)B + theta(2)B + ...). True signs.}

\item{bphi}{coefficients of the seasonal auto-regressive polynomial. True signs.}

\item{bd}{seasonal differencing order.}

\item{btheta}{coefficients of the seasonal moving average polynomial. True signs.}
}
\value{
a `"JD3_SARIMA"` model.
}
\description{
Seasonal ARIMA model (Box-Jenkins)
}
