% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arima.R
\name{sarima.properties}
\alias{sarima.properties}
\title{SARIMA Properties}
\usage{
sarima.properties(model, nspectrum = 601, nacf = 36)
}
\arguments{
\item{model}{a `"JD3_SARIMA"` model (created with [sarima.model()]).}

\item{nspectrum}{number of points in \[0, pi\] to calculate the spectrum.}

\item{nacf}{maximum lag at which to calculate the acf.}
}
\description{
SARIMA Properties
}
\examples{
mod1 <- sarima.model(period = 12, d =1, bd = 1, theta = 0.2, btheta = 0.2)
sarima.properties(mod1)
}
