% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arima.R
\name{sarima_decompose}
\alias{sarima_decompose}
\title{Decompose SARIMA Model}
\usage{
sarima_decompose(model, rmod = 0, epsphi = 0)
}
\arguments{
\item{model}{SARIMA model to decompose.}

\item{rmod}{trend threshold.}

\item{epsphi}{seasonal tolerance (in degrees).}
}
\description{
Decompose SARIMA Model
}
\examples{
model <- sarima_model(period = 12, d =1, bd = 1, theta = -0.6, btheta = -0.5)
ucm <- sarima_decompose(model)

}
\seealso{
Other SARIMA functions: 
\code{\link{arima_model}()},
\code{\link{arima_properties}()},
\code{\link{arima_sum}()},
\code{\link{sarima_estimate}()},
\code{\link{sarima_model}()},
\code{\link{sarima_properties}()},
\code{\link{sarima_random}()},
\code{\link{ucarima_canonical}()},
\code{\link{ucarima_estimate}()},
\code{\link{ucarima_model}()},
\code{\link{ucarima_wk}()}
}
\concept{SARIMA functions}
