% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tests_seasonality.R
\name{seasonality.combined}
\alias{seasonality.combined}
\title{"X12" Test On Seasonality}
\usage{
seasonality.combined(data, period, firstperiod = 1, mul = TRUE)
}
\arguments{
\item{data}{the input data.}

\item{period}{Tested periodicity.}

\item{firstperiod}{First period of the series (1-based)}

\item{mul}{boolean indicating if the seasonal decomposition is multiplicative (`mul = TRUE`) or additive (`mul = FALSE`).}
}
\description{
"X12" Test On Seasonality
}
\details{
Combined test on the presence of identifiable seasonality (see Ladiray and Quenneville, 1999).
}
\examples{
seasonality.combined(ABS$X0.2.09.10.M, 12)
seasonality.combined(randomsT(2, 1000), 7)
}
