% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tests_seasonality.R
\name{seasonality.friedman}
\alias{seasonality.friedman}
\title{Friedman Seasonality Test}
\usage{
seasonality.friedman(data, period, nyears = 0)
}
\arguments{
\item{data}{the input data.}

\item{period}{Tested periodicity.}

\item{nyears}{Number of number of periods number of cycles considered in the test, at the end of the series:
in periods (positive value) or years (negative values).
By default (\code{nyears = 0}), the entire sample is used.}
}
\value{
A `c("JD3_TEST", "JD3")` object (see [statisticaltest()] for details).
}
\description{
Friedman Seasonality Test
}
\details{
Non parametric test ("ANOVA"-type).
}
