% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tests_seasonality.R
\name{seasonality.periodogram}
\alias{seasonality.periodogram}
\title{Periodogram Seasonality Test}
\usage{
seasonality.periodogram(data, period, nyears = 0)
}
\arguments{
\item{data}{the input data.}

\item{period}{Tested periodicity.}

\item{nyears}{Number of number of periods number of cycles considered in the test, at the end of the series:
in periods (positive value) or years (negative values).
By default (\code{nyears = 0}), the entire sample is used.}
}
\value{
A `c("JD3_TEST", "JD3")` object (see [statisticaltest()] for details).
}
\description{
Periodogram Seasonality Test
}
\details{
Tests on the sum of a periodogram at seasonal frequencies.
}
\examples{
seasonality.periodogram(ABS$X0.2.09.10.M, 12)
seasonality.periodogram(randomsT(2, 1000), 7)
}
