% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tests_seasonality.R
\name{seasonality_canovahansen}
\alias{seasonality_canovahansen}
\title{Seasonal Canova-Hansen test}
\usage{
seasonality_canovahansen(data, p0, p1, np, original = FALSE)
}
\arguments{
\item{data}{the input data.}

\item{p0}{Initial periodicity (included)}

\item{p1}{Final periodicity (included)}

\item{np}{Number of periodicities equally spaced in $[p0,p1]$}

\item{original}{\code{TRUE} for original algorithm, False for solution proposed by T. Proietti (based on Ox code)}
}
\description{
Seasonal Canova-Hansen test
}
\seealso{
Other seasonality tests: 
\code{\link{seasonality_combined}()},
\code{\link{seasonality_friedman}()},
\code{\link{seasonality_f}()},
\code{\link{seasonality_kruskalwallis}()},
\code{\link{seasonality_periodogram}()},
\code{\link{seasonality_qs}()}
}
\concept{seasonality tests}
