% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tests_seasonality.R
\name{seasonality_combined}
\alias{seasonality_combined}
\title{"X12" Test On Seasonality}
\usage{
seasonality_combined(data, period, firstperiod = 1, mul = TRUE)
}
\arguments{
\item{data}{the input data.}

\item{period}{Tested periodicity.
If \code{data} is a \code{ts} object, by default \code{period} is set to the frequency of \code{data}.}

\item{firstperiod}{First period of the series (1-based)}

\item{mul}{boolean indicating if the seasonal decomposition is multiplicative (\code{mul = TRUE}) or additive (\code{mul = FALSE}).}
}
\description{
"X12" Test On Seasonality
}
\details{
Combined test on the presence of identifiable seasonality (see Ladiray and Quenneville, 1999).
}
\examples{
seasonality_combined(ABS$X0.2.09.10.M, 12)
seasonality_combined(randoms_t(2, 1000), 7)
}
\seealso{
Other seasonality tests: 
\code{\link{seasonality_canovahansen}()},
\code{\link{seasonality_friedman}()},
\code{\link{seasonality_f}()},
\code{\link{seasonality_kruskalwallis}()},
\code{\link{seasonality_periodogram}()},
\code{\link{seasonality_qs}()}
}
\concept{seasonality tests}
