% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tests_seasonality.R
\name{seasonality_friedman}
\alias{seasonality_friedman}
\title{Friedman Seasonality Test}
\usage{
seasonality_friedman(data, period, nyears = 0)
}
\arguments{
\item{data}{the input data.}

\item{period}{Tested periodicity.
If \code{data} is a \code{ts} object, by default \code{period} is set to the frequency of \code{data}.}

\item{nyears}{Number of number of periods number of cycles considered in the test, at the end of the series:
in periods (positive value) or years (negative values).
By default (\code{nyears = 0}), the entire sample is used.}
}
\value{
A \code{c("JD3_TEST", "JD3")} object (see \code{\link[=statisticaltest]{statisticaltest()}} for details).
}
\description{
Friedman Seasonality Test
}
\details{
Non parametric test ("ANOVA"-type).
}
\examples{
seasonality_friedman(ABS$X0.2.09.10.M, 12)
}
\seealso{
Other seasonality tests: 
\code{\link{seasonality_canovahansen}()},
\code{\link{seasonality_combined}()},
\code{\link{seasonality_f}()},
\code{\link{seasonality_kruskalwallis}()},
\code{\link{seasonality_periodogram}()},
\code{\link{seasonality_qs}()}
}
\concept{seasonality tests}
