% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec_regarima.R
\name{set_arima}
\alias{set_arima}
\title{Set ARIMA Specification}
\usage{
set_arima(
  x,
  mean = NA,
  mean.type = c(NA, "Undefined", "Fixed", "Initial"),
  p = NA,
  d = NA,
  q = NA,
  bp = NA,
  bd = NA,
  bq = NA,
  coef = NA,
  coef.type = c(NA, "Undefined", "Fixed", "Initial")
)
}
\arguments{
\item{x}{the specification.}

\item{mean}{to set the coefficient of the mean. If \code{mean = 0}, the mean is disabled.}

\item{mean.type}{a character defining the mean coefficient estimation procedure.
Possible procedures are: \code{"Undefined"} = no use of any user-defined input (i.e. coefficients are estimated),
\code{"Fixed"} = the coefficients are fixed at the value provided by the user,
\code{"Initial"} = the value defined by the user is used as the initial condition.}

\item{p, d, q, bp, bd, bq}{to specify the order of the SARIMA model in the form ARIMA(p,d,q)(bp,bd,bd).}

\item{coef}{a vector providing the coefficients for the regular and seasonal AR and MA polynominals.
The vector length must be equal to the sum of the regular and seasonal AR and MA orders.
The coefficients shall be provided in the following order: regular AR (\emph{Phi}; \code{p} elements),
regular MA  (\emph{Theta}; \code{q} elements), seasonal AR (\emph{BPhi}; \code{bp} elements)
and seasonal MA (\emph{BTheta}; \code{bq} elements).
E.g.: \code{arima.coef=c(0.6,0.7)} with \code{p=1, q=0,bp=1} and \code{bq=0}.}

\item{coef.type}{a vector defining the ARMA coefficients estimation procedure.
Possible procedures are: \code{"Undefined"} = no use of any user-defined input (i.e. coefficients are estimated),
\code{"Fixed"} = the coefficients are fixed at the value provided by the user,
\code{"Initial"} = the value defined by the user is used as the initial condition.}
}
\description{
Function to set the ARIMA model when the automatic modelling is disabled.
}
