% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec_regarima.R
\name{set_basic}
\alias{set_basic}
\title{Set Basic Specification}
\usage{
set_basic(
  x,
  type = c(NA, "All", "From", "To", "Between", "Last", "First", "Excluding"),
  d0 = NULL,
  d1 = NULL,
  n0 = 0,
  n1 = 0,
  preliminary.check = NA,
  preprocessing = NA
)
}
\arguments{
\item{x}{the specification.}

\item{type, d0, d1, n0, n1}{parameters to specifiy the span of to be used.

\code{d0} and \code{d1} characters in the format "YYYY-MM-DD" to specify first/last date of the span when \code{type} equals to \code{"From"}, \code{"To"} or \code{"Between"}.

\code{n0} and \code{n1} numerics to specify the number of periods at the beginning/end of the series to be used for the span
(\code{type} equals to \code{"From"}, \code{"To"}) or to exclude (\code{type} equals to \code{"Excluding"}).}

\item{preliminary.check}{a boolean to check the quality of the input series and exclude highly problematic ones
(e.g. the series with a number of identical observations and/or missing values above pre-specified threshold values).}

\item{preprocessing}{(REGARIMA/X13 Specific) a boolean to enable/disable the pre-processing.}
}
\description{
Set Basic Specification
}
