% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec_benchmarking.R
\name{set_benchmarking}
\alias{set_benchmarking}
\title{Set Benchmarking Specification}
\usage{
set_benchmarking(
  x,
  enabled = NA,
  target = c(NA, "CalendarAdjusted", "Normal"),
  rho = NA,
  lambda = NA,
  forecast = NA,
  bias = c(NA, "None")
)
}
\arguments{
\item{x}{the specification.}

\item{enabled}{Boolean to enable the user to perform benchmarking.}

\item{target}{specifies the target variable for the benchmarking procedure, which can be the raw series (\code{"Original"}); or the series adjusted for calendar effects (\code{"CalendarAdjusted"}).}

\item{rho}{the value of the AR(1) parameter (set between 0 and 1) in the function used for benchmarking.}

\item{lambda}{a parameter in the function used for benchmarking that relates to the weights in the regression equation; it is typically equal to 0, 1/2 or 1.}

\item{forecast}{Boolean indicating if the forecasts of the seasonally adjusted series and of the target variable (\code{target}) are used in the benchmarking computation so the benchmarking constrain is applied also to the forecasting period.}

\item{bias}{TODO}
}
\description{
Set Benchmarking Specification
}
