% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec_regarima.R
\name{set_easter}
\alias{set_easter}
\title{Set Easter Specification}
\usage{
set_easter(
  x,
  enabled = NA,
  julian = NA,
  duration = NA,
  test = c(NA, "Add", "Remove", "None"),
  coef = NA,
  coef.type = c(NA, "Estimated", "Fixed"),
  type = c(NA, "Unused", "Standard", "IncludeEaster", "IncludeEasterMonday")
)
}
\arguments{
\item{x}{the specification.}

\item{enabled}{a logical indicating if the program considers the Easter effect in the model.}

\item{julian}{a logical indicating if the program uses the Julian Easter (expressed in Gregorian calendar).}

\item{duration}{a numeric indicating the duration of the Easter effect (length in days, between 1 and 20).}

\item{test}{defines the pre-tests for the significance of the Easter effect based on the t-statistic
(the Easter effect is considered as significant if the t-statistic is greater than 1.96):
\code{"Add"} = the Easter effect variable is not included in the initial regression model but can be added
to the RegARIMA model after the test;
\code{"Remove"} = the Easter effect variable belongs to the initial regression model but can be removed
from the RegARIMA model after the test;
\code{"None"} = the Easter effect variable is not pre-tested and is included in the model.}

\item{coef}{to set the coefficient of the easter regressor.}

\item{coef.type}{a character defining the easter regressor coefficient estimation procedure.
Possible procedures are: \code{"Estimated"} = coefficient is estimated,
\code{"Fixed"} = the coefficients is fixed.}

\item{type}{(TRAMO specific) a \code{character} that specifies the presence and the length of the Easter effect:
\code{"Unused"} = the Easter effect is not considered; \code{"Standard"} = influences the period of \code{n} days strictly before Easter Sunday;
\code{"IncludeEaster"} = influences the entire period (\code{n}) up to and including Easter Sunday;
\code{"IncludeEasterMonday"} = influences the entire period (\code{n}) up to and including Easter Monday.}
}
\description{
Set Easter Specification
}
