% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec_regarima.R
\name{set_estimate}
\alias{set_estimate}
\title{Set Estimate Specification}
\usage{
set_estimate(
  x,
  type = c(NA, "All", "From", "To", "Between", "Last", "First", "Excluding"),
  d0 = NULL,
  d1 = NULL,
  n0 = 0,
  n1 = 0,
  tol = NA,
  exact.ml = NA,
  unit.root.limit = NA
)
}
\arguments{
\item{x}{the specification.}

\item{type, d0, d1, n0, n1}{parameters to specifiy the span of to be used.

\code{d0} and \code{d1} characters in the format "YYYY-MM-DD" to specify first/last date of the span when \code{type} equals to \code{"From"}, \code{"To"} or \code{"Between"}.

\code{n0} and \code{n1} numerics to specify the number of periods at the beginning/end of the series to be used for the span
(\code{type} equals to \code{"From"}, \code{"To"}) or to exclude (\code{type} equals to \code{"Excluding"}).}

\item{tol}{a numeric, convergence tolerance. The absolute changes in the log-likelihood function
are compared to this value to check for the convergence of the estimation iterations.}

\item{exact.ml}{(TRAMO specific) \code{logical}, the exact maximum likelihood estimation. If \code{TRUE}, the program performs an exact maximum likelihood estimation.
If \code{FASLE}, the Unconditional Least Squares method is used.}

\item{unit.root.limit}{(TRAMO specific) \code{numeric}, the final unit root limit. The threshold value for the final unit root test
for identification of differencing orders. If the magnitude of an AR root for the final model is smaller than this number,
then a unit root is assumed, the order of the AR polynomial is reduced by one and the appropriate order of the differencing
(non-seasonal, seasonal) is increased.}
}
\description{
Set Estimate Specification
}
