% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec_regarima.R
\name{set_outlier}
\alias{set_outlier}
\title{Set Outlier Specification}
\usage{
set_outlier(
  x,
  span.type = c(NA, "All", "From", "To", "Between", "Last", "First", "Excluding"),
  d0 = NULL,
  d1 = NULL,
  n0 = 0,
  n1 = 0,
  outliers.type = NA,
  critical.value = NA,
  tc.rate = NA,
  method = c(NA, "AddOne", "AddAll"),
  maxiter = NA,
  lsrun = NA,
  eml.est = NA
)
}
\arguments{
\item{x}{the specification.}

\item{span.type, d0, d1, n0, n1}{parameters to specifiy the span of to be used.

\code{d0} and \code{d1} characters in the format "YYYY-MM-DD" to specify first/last date of the span when \code{type} equals to \code{"From"}, \code{"To"} or \code{"Between"}.

\code{n0} and \code{n1} numerics to specify the number of periods at the beginning/end of the series to be used for the span
(\code{type} equals to \code{"From"}, \code{"To"}) or to exclude (\code{type} equals to \code{"Excluding"}).}

\item{outliers.type}{vector of characters of the outliers to be automatically detected. \code{"AO"} for additive outliers, \code{"TC"} for transitory changes
\code{"LS"} for level shifts and \code{"SO"} for seasonal outliers.
For example \code{outliers.type = c("AO", "LS")} to enable the detection of additive outliers and level shifts.
If \code{outliers.type = NULL} or \code{outliers.type = character()}, automatic detection of outliers is disabled.}

\item{critical.value}{\code{numeric}. The entered critical value for the outliers' detection procedure.
If equal to 0 the critical value for the outliers' detection procedure is automatically determined
by the number of observations in the outlier detection time span.}

\item{tc.rate}{the rate of decay for the transitory change outlier (by default to \code{0.7}).}

\item{method}{(REGARIMA/X13 Specific) determines how the program successively adds detected outliers to the model.
At present, only the \code{"AddOne"} method is supported.}

\item{maxiter}{(REGARIMA/X13 Specific) maximum number of iterations (by default to \code{30}).}

\item{lsrun}{(REGARIMA/X13 Specific) number of successive level shifts to test for cancellation (by default to \code{0}).}

\item{eml.est}{(TRAMO Specific) \code{logical} for the exact likelihood estimation method. It controls the method applied for a parameter estimation
in the intermediate steps of the automatic detection and correction of outliers. If \code{TRUE}, an exact likelihood estimation method is used.
When \code{FALSE}, the fast Hannan-Rissanen method is used.}
}
\description{
Set Outlier Specification
}
