% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec_regarima.R
\name{set_tradingdays}
\alias{set_tradingdays}
\title{Set Trading-Days Specification}
\usage{
set_tradingdays(
  x,
  option = c(NA, "TradingDays", "WorkingDays", "TD3", "TD3c", "TD4", "None",
    "UserDefined"),
  uservariable = NA,
  stocktd = NA,
  test = c(NA, "None", "Remove", "Add", "Separate_T", "Joint_F"),
  coef = NA,
  coef.type = c(NA, "Fixed", "Estimated"),
  automatic = c(NA, "Unused", "FTest", "WaldTest", "Aic", "Bic"),
  pftd = NA,
  autoadjust = NA,
  leapyear = c(NA, "LeapYear", "LengthOfPeriod", "None"),
  leapyear.coef = NA,
  leapyear.coef.type = c(NA, "Fixed", "Estimated")
)
}
\arguments{
\item{x}{the specification.}

\item{option}{to specify the set of trading days regression variables:
\code{"TradingDays"} = six day-of-the-week regression variables;
\code{"WorkingDays"} = one working/non-working day contrast variable;
\code{"TD3"} = weeks, Saturdays, Sundays;
\code{"TD3c"} = weeks, Fridays+Saturdays, Sundays;
\code{"TD4"} = weeks, Fridays, Saturdays, Sundays;
\code{"None"} = no correction for trading days and working days effects;
\code{"UserDefined"} = userdefined trading days regressors.}

\item{uservariable}{a vector of characters to specify the name of user-defined calendar regressors.
When specified, automatically set \code{option = "UserDefined"}.}

\item{stocktd}{a numeric indicating the day of the month when inventories and other stock are reported
(to denote the last day of the month, set the variable to 31).
When specified, automatically set \code{option = "None"}.}

\item{test}{defines the pre-tests for the significance of the trading day regression variables
based on the AICC statistics: \code{"None"} = the trading day variables are not pre-tested and are included in the model;

(REGARIMA/X-13 specific)

\code{"Add"} = the trading day variables are not included in the initial regression model
but can be added to the RegARIMA model after the test;
\code{"Remove"} = the trading day variables belong to the initial regression model but can be removed from the RegARIMA model
after the test;

(TRAMO specific)

\code{"Separate_T"} = a t-test is applied to each trading day variable separately and the trading day variables are included in the RegArima model
if at least one t-statistic is greater than 2.6 or if two t-statistics are greater than 2.0 (in absolute terms);
\code{"Joint_F"} = a joint F-test of significance of all the trading day variables. The trading day effect is significant if the F statistic is greater than 0.95.}

\item{coef}{vector of coefficients for the tranding-days regressors.}

\item{coef.type, leapyear.coef.type}{vector defining if the coefficients are fixed or estimated.}

\item{automatic}{defines whether the calendar effects should be added to the model manually (\code{"Unused"}) or automatically.
During the automatic selection, the choice of the number of calendar variables can be based on the F-Test (\code{"FTest"}, TRAMO specific), the Wald Test (\code{"WaldTest"}), or by minimising AIC or BIC;
the model with higher F value is chosen, provided that it is higher than \code{pftd}).}

\item{pftd}{(TRAMO SPECIFIC) \code{numeric}. The p-value used to assess the significance of the pre-tested calendar effects.}

\item{autoadjust}{a logical indicating if the program corrects automatically for the leap year effect if the leap year regressors is significant. Only used when the data is log transformed.}

\item{leapyear}{a \code{character} to specify whether or not to include the leap-year effect in the model:
\code{"LeapYear"} = leap year effect; \code{"LengthOfPeriod"} = length of period (REGARIMA/X-13 specific), \code{"None"} = no effect included.}

\item{leapyear.coef}{coefficient of the leapyear regressor.}
}
\description{
Set Trading-Days Specification
}
