% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calendars.R
\name{single_day}
\alias{single_day}
\title{Single Date}
\usage{
single_day(date, weight = 1)
}
\arguments{
\item{date}{the date of the holiday in the format \code{"YYYY-MM-DD"}.}

\item{weight}{weight associated to the holiday.}
}
\description{
Single Date
}
\examples{
single_day("1999-03-19")
}
\seealso{
Other calendar functions: 
\code{\link{calendar_td}()},
\code{\link{easter_dates}()},
\code{\link{easter_day}()},
\code{\link{fixed_day}()},
\code{\link{fixed_week_day}()},
\code{\link{holidays}()},
\code{\link{long_term_mean}()},
\code{\link{national_calendar}()},
\code{\link{special_day}()},
\code{\link{stock_td}()},
\code{\link{td}()}
}
\concept{calendar functions}
