% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calendars.R
\name{stock_td}
\alias{stock_td}
\title{Stock Trading days}
\usage{
stock_td(frequency, start, length, s, w = 31)
}
\arguments{
\item{frequency}{Frequency of the series, number of periods per year (12,4,3,2..)}

\item{start, length}{First date (array with the first year and the first period)
(for instance \code{c(1980, 1)}) and number of periods of the output variables. Can also be provided with the \code{s} argument}

\item{s}{time series used to get the dates for the trading days variables. If supplied the
parameters \code{frequency}, \code{start} and \code{length} are ignored.}

\item{w}{indicates day of the month when inventories and other stocks are reported
(to denote the last day of the month enter 31).}
}
\description{
Stock Trading days
}
