% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tests.R
\name{testofruns}
\alias{testofruns}
\title{Runs test around the mean or the median}
\usage{
testofruns(data, mean = T, number = T)
}
\arguments{
\item{data}{Data being tested}

\item{mean}{If True, runs around the mean. Otherwise, runs around the median}

\item{number}{If True, test the number of runs. Otherwise, test the lengths of the runs}
}
\value{

}
\description{
Runs test around the mean or the median
}
