% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tests.R
\name{testofupdownruns}
\alias{testofupdownruns}
\title{Up and down runs test}
\usage{
testofupdownruns(data, number = T)
}
\arguments{
\item{number}{If True, test the number of runs. Otherwise, test the lengths of the runs}
}
\value{

}
\description{
Up and down runs test
}
