% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arima.R
\name{ucarima_canonical}
\alias{ucarima_canonical}
\title{Title}
\usage{
ucarima_canonical(ucm, cmp = 0, adjust = TRUE)
}
\arguments{
\item{ucm}{UCARIMA model returned by \code{\link[=ucarima_model]{ucarima_model()}}.}

\item{adjust}{}
}
\description{
Title
}
\seealso{
Other SARIMA functions: 
\code{\link{arima_model}()},
\code{\link{arima_properties}()},
\code{\link{arima_sum}()},
\code{\link{sarima_decompose}()},
\code{\link{sarima_estimate}()},
\code{\link{sarima_model}()},
\code{\link{sarima_properties}()},
\code{\link{sarima_random}()},
\code{\link{ucarima_estimate}()},
\code{\link{ucarima_model}()},
\code{\link{ucarima_wk}()}
}
\concept{SARIMA functions}
