% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arima.R
\name{ucarima_estimate}
\alias{ucarima_estimate}
\title{Estimate UCARIMA Model}
\usage{
ucarima_estimate(x, ucm, stdev = TRUE)
}
\arguments{
\item{x}{univariate time series}

\item{ucm}{UCARIMA model returned by \code{\link[=ucarima_model]{ucarima_model()}}.}

\item{stdev}{}
}
\value{
matrix containing the different components.
}
\description{
Estimate UCARIMA Model
}
\seealso{
Other SARIMA functions: 
\code{\link{arima_model}()},
\code{\link{arima_properties}()},
\code{\link{arima_sum}()},
\code{\link{sarima_decompose}()},
\code{\link{sarima_estimate}()},
\code{\link{sarima_model}()},
\code{\link{sarima_properties}()},
\code{\link{sarima_random}()},
\code{\link{ucarima_canonical}()},
\code{\link{ucarima_model}()},
\code{\link{ucarima_wk}()}
}
\concept{SARIMA functions}
