% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arima.R
\name{ucarima_model}
\alias{ucarima_model}
\title{Title}
\usage{
ucarima_model(model = NULL, components, complements = NULL, checkmodel = F)
}
\arguments{
\item{complements}{Complements of (some) components}
}
\description{
Title
}
\seealso{
Other SARIMA functions: 
\code{\link{arima_model}()},
\code{\link{arima_properties}()},
\code{\link{arima_sum}()},
\code{\link{sarima_decompose}()},
\code{\link{sarima_estimate}()},
\code{\link{sarima_model}()},
\code{\link{sarima_properties}()},
\code{\link{sarima_random}()},
\code{\link{ucarima_canonical}()},
\code{\link{ucarima_estimate}()},
\code{\link{ucarima_wk}()}
}
\concept{SARIMA functions}
