% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arima.R
\name{ucarima_wk}
\alias{ucarima_wk}
\title{Wiener Kolmogorov Estimators}
\usage{
ucarima_wk(ucm, cmp, signal = TRUE, nspectrum = 601, nwk = 300)
}
\arguments{
\item{ucm}{UCARIMA model returned by \code{\link[=ucarima_model]{ucarima_model()}}.}

\item{nwk}{}
}
\description{
Wiener Kolmogorov Estimators
}
\seealso{
Other SARIMA functions: 
\code{\link{arima_model}()},
\code{\link{arima_properties}()},
\code{\link{arima_sum}()},
\code{\link{sarima_decompose}()},
\code{\link{sarima_estimate}()},
\code{\link{sarima_model}()},
\code{\link{sarima_properties}()},
\code{\link{sarima_random}()},
\code{\link{ucarima_canonical}()},
\code{\link{ucarima_estimate}()},
\code{\link{ucarima_model}()}
}
\concept{SARIMA functions}
