% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calendars.R
\name{weighted_calendar}
\alias{weighted_calendar}
\title{Composite Calendars}
\usage{
weighted_calendar(calendars, weights)
}
\arguments{
\item{calendars}{list of calendars.}

\item{weights}{vector weights associated to each calendar.}
}
\description{
Composite Calendars
}
\details{
Composite calendars are useful for series that includes data from more than one country/region.
They can be used, for example, to create the calendar for the European Union or to create the national calendar for a country,
in which regional holidays are celebrated.
For example, in Germany public holidays are determined by the federal states.
Therefore, Epiphany is celebrated only in Baden-Wurttemberg, Bavaria and in Saxony-Anhalt, while from 1994 Day of Repentance and Prayer is celebrated only in Saxony.
}
