% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec_regarima.R
\name{add_outlier}
\alias{add_outlier}
\alias{remove_outlier}
\alias{add_ramp}
\alias{remove_ramp}
\title{Manage Outliers/Ramps in Specification}
\usage{
add_outlier(x, type, date, name = sprintf("\%s (\%s)", type, date), coef = 0)

remove_outlier(x, type = NULL, date = NULL, name = NULL)

add_ramp(x, start, end, name = sprintf("rp.\%s - \%s", start, end), coef = 0)

remove_ramp(x, start = NULL, end = NULL, name = NULL)
}
\arguments{
\item{x}{the specification.}

\item{type, }{date type and date of the outliers. Possible \code{type} are:
\code{"AO"} = additive, \code{"LS"} = level shift, \code{"TC"} = transitory change and
\code{"SO"} = seasonal outlier.}

\item{name}{the name of the variable (to format print).}

\item{coef}{the coefficient. If equal to 0 the outliers/ramps are estimated.}

\item{start, end}{dates of the ramp regressor.}
}
\description{
Generic function to add outliers or Ramp regressors (\code{add_outlier()} and \code{add_ramp()})
to a specification or to remove them (\code{remove_outlier()} and \code{remove_ramp()}).
}
