% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec_regarima.R
\name{add_usrdefvar}
\alias{add_usrdefvar}
\title{Add User-Defined Variable}
\usage{
add_usrdefvar(
  x,
  id,
  name = NULL,
  lag = 0,
  coef = NULL,
  regeffect = c("Undefined", "Trend", "Seasonal", "Irregular", "Series",
    "SeasonallyAdjusted")
)
}
\arguments{
\item{x}{the specification.}

\item{id}{the id of the variable in the format \code{"group_name.name"}.}

\item{name}{name the name of the variable (to format print). By default equals to \code{id}.}

\item{lag}{integer defining if the user-defined variable should be lagged.
By default (\code{lag = 0}), the regressor \eqn{x_t} is not lagged. If \code{lag = 1}, then \eqn{x_{t-1}} is used.}

\item{coef}{the coefficient.}

\item{regeffect}{assigned component of the user-defined variable.
By default (\code{"Undefined"}), the variable is assigned to an additional component: the variable is used to improve the modelling, but it is not removed from the series for the decomposition.}
}
\description{
Add user-defined variable to a specification. To add user-defined calendar regressors, see \code{\link[=set_tradingdays]{set_tradingdays()}}.
}
