% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calendars.R
\name{easter_dates}
\alias{easter_dates}
\title{Compute Easter days between two years}
\usage{
easter_dates(year0, year1, julian = FALSE)
}
\arguments{
\item{year0, year1}{years.}

\item{julian}{boolean indicating if julian calendar must be used.}
}
\description{
Compute Easter days between two years
}
\examples{
easter_dates(2020, 2021)
}
