% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calendars.R
\name{easter_day}
\alias{easter_day}
\title{Easter Related Day}
\usage{
easter_day(offset, julian = FALSE, weight = 1, validity = NULL)
}
\arguments{
\item{offset}{The position of the holiday in relation to the Easter Sunday, measured in days (can be positive or negative).}

\item{julian}{boolean indicating if julian calendar must be used.}

\item{weight}{weight associated to the holiday.}

\item{validity}{validity period: either `NULL` (full sample) or a named list with `"start"` and/or "end" dates in the format `"YYYY-MM-DD"`.}
}
\description{
Easter Related Day
}
\examples{
easter_day(1) #Easter Monday
easter_day(-2) # Easter Good Friday

}
