% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calendars.R
\name{fixed_week_day}
\alias{fixed_week_day}
\title{Fixed Week Day}
\usage{
fixed_week_day(month, week, dayofweek, weight = 1, validity = NULL)
}
\arguments{
\item{month}{month of the holiday: from \code{1} (January) to \code{12} (December).}

\item{week}{position of the specified week day in the month: from \code{1} (first week of the month) to \code{5}. Should be always lower than 5.
\code{-1} for the last \code{dayofweek} of the month.}

\item{dayofweek}{day of the week: from \code{1} (Monday) to \code{7} (Sunday).}

\item{weight}{weight associated to the holiday.}

\item{validity}{validity period: either \code{NULL} (full sample) or a named list with \code{"start"} and/or "end" dates in the format \code{"YYYY-MM-DD"}.}
}
\description{
Fixed Week Day
}
\examples{
day <- fixed_week_day(7, 2, 3) # second Tuesday of July
day

}
