% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jd2r.R, R/protobuf.R, R/arima.R, R/jd3rslts.R,
%   R/modellingcontext.R, R/procresults.R, R/regarima_rslts.R, R/zzz.R
\docType{data}
\name{ts_r2jd}
\alias{ts_r2jd}
\alias{tsdomain_r2jd}
\alias{ts_jd2r}
\alias{matrix_jd2r}
\alias{matrix_r2jd}
\alias{j2r_ldt}
\alias{j2r_dt}
\alias{r2j_dt}
\alias{r2j_ldt}
\alias{jd2r_parameters}
\alias{jdomain}
\alias{jd3_utilities}
\alias{enum_sextract}
\alias{enum_sof}
\alias{enum_extract}
\alias{enum_of}
\alias{r2p_parameter}
\alias{p2r_parameter}
\alias{r2p_parameters}
\alias{r2p_lparameters}
\alias{r2p_tparameters}
\alias{p2r_parameters}
\alias{p2r_parameters_rslt}
\alias{p2r_parameters_rsltx}
\alias{p2r_test}
\alias{p2r_matrix}
\alias{r2p_matrix}
\alias{p2r_ts}
\alias{r2p_ts}
\alias{p2r_parameters_estimation}
\alias{p2r_likelihood}
\alias{p2r_date}
\alias{r2p_date}
\alias{p2r_span}
\alias{r2p_span}
\alias{p2r_arima}
\alias{p2r_ucarima}
\alias{p2r_spec_sarima}
\alias{r2p_spec_sarima}
\alias{p2r_outliers}
\alias{r2p_outliers}
\alias{p2r_ramps}
\alias{r2p_ramps}
\alias{p2r_uservars}
\alias{r2p_uservars}
\alias{p2r_variables}
\alias{p2r_sa_component}
\alias{p2r_sa_decomposition}
\alias{p2r_sa_diagnostics}
\alias{p2r_spec_benchmarking}
\alias{r2p_spec_benchmarking}
\alias{r2jd_sarima}
\alias{jd2r_ucarima}
\alias{proc_numeric}
\alias{proc_vector}
\alias{proc_int}
\alias{proc_bool}
\alias{proc_ts}
\alias{proc_str}
\alias{proc_desc}
\alias{proc_test}
\alias{proc_parameter}
\alias{proc_parameters}
\alias{proc_matrix}
\alias{proc_data}
\alias{proc_dictionary}
\alias{proc_dictionary2}
\alias{proc_likelihood}
\alias{.r2p_moniker}
\alias{.p2r_moniker}
\alias{.r2p_datasupplier}
\alias{.p2r_datasupplier}
\alias{.r2p_datasuppliers}
\alias{.p2r_datasuppliers}
\alias{.p2r_context}
\alias{.r2p_context}
\alias{.p2jd_context}
\alias{.jd2p_context}
\alias{context2dict}
\alias{.jd2r_modellingcontext}
\alias{.r2jd_modellingcontext}
\alias{jd3Object}
\alias{p2r_regarima_rslts}
\alias{DATE_MIN}
\alias{DATE_MAX}
\title{Java Utility Functions}
\format{
An object of class \code{Message} of length 3.

An object of class \code{Message} of length 3.
}
\usage{
ts_r2jd(s)

tsdomain_r2jd(period, startYear, startPeriod, length)

ts_jd2r(s)

matrix_jd2r(s)

matrix_r2jd(s)

j2r_ldt(ldt)

j2r_dt(dt)

r2j_dt(dt)

r2j_ldt(dt)

jd2r_parameters(jparams)

jdomain(period, start, end)

enum_sextract(type, p)

enum_sof(type, code)

enum_extract(type, p)

enum_of(type, code, prefix)

r2p_parameter(r)

p2r_parameter(p)

r2p_parameters(r)

r2p_lparameters(r)

r2p_tparameters(data, type)

p2r_parameters(p)

p2r_parameters_rslt(p)

p2r_parameters_rsltx(p)

p2r_test(p)

p2r_matrix(p)

r2p_matrix(r)

p2r_ts(p)

r2p_ts(r)

p2r_parameters_estimation(p)

p2r_likelihood(p)

p2r_date(p)

r2p_date(s)

p2r_span(span)

r2p_span(rspan)

p2r_test(p)

p2r_arima(p)

p2r_ucarima(p)

p2r_spec_sarima(spec)

r2p_spec_sarima(r)

p2r_outliers(p)

r2p_outliers(r)

p2r_ramps(p)

r2p_ramps(r)

p2r_uservars(p)

r2p_uservars(r)

p2r_variables(p)

p2r_sa_component(p)

p2r_sa_decomposition(p, full = F)

p2r_sa_diagnostics(p)

p2r_spec_benchmarking(p)

r2p_spec_benchmarking(r)

r2jd_sarima(model)

jd2r_ucarima(jucm)

proc_numeric(rslt, name)

proc_vector(rslt, name)

proc_int(rslt, name)

proc_bool(rslt, name)

proc_ts(rslt, name)

proc_str(rslt, name)

proc_desc(rslt, name)

proc_test(rslt, name)

proc_parameter(rslt, name)

proc_parameters(rslt, name)

proc_matrix(rslt, name)

proc_data(rslt, name)

proc_dictionary(name)

proc_dictionary2(jobj)

proc_likelihood(jrslt, prefix)

.r2p_moniker(r)

.p2r_moniker(p)

.r2p_datasupplier(name, r)

.p2r_datasupplier(p)

.r2p_datasuppliers(r)

.p2r_datasuppliers(p)

.p2r_context(p)

.r2p_context(r)

.p2jd_context(p)

.jd2p_context(jd)

context2dict(x)

.jd2r_modellingcontext(jcontext)

.r2jd_modellingcontext(r)

jd3Object(jobjRef, subclasses = NULL, result = F)

p2r_regarima_rslts(p)

DATE_MIN

DATE_MAX
}
\arguments{
\item{p, r, spec, model, jucm, vp, start, end, name, x}{parameters.}
}
\description{
These functions are used in all JDemetra+ 3.0 packages to easily interact between R and Java objects.
}
\keyword{datasets}
