% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arima.R
\name{sarima_random}
\alias{sarima_random}
\title{Simulate Seasonal ARIMA}
\usage{
sarima_random(model, length, stde = 1, tdegree = 0)
}
\arguments{
\item{model}{a `"JD3_SARIMA"` model (see [sarima_model()] function).}

\item{length}{length of the output series.}

\item{stde}{the standard deviation of the normal distribution of the innovations of the simulated series. Unused if tdegree is larger than 0.}

\item{tdegree}{Degrees of freedom of the T distribution of the innovations. O if normal distribution is used.}
}
\description{
Simulate Seasonal ARIMA
}
\examples{
# Airline model
s_model <- sarima_model(period = 12, d =1, bd = 1, theta = 0.2, btheta = 0.2)
x <- sarima_random(s_model, length = 64)
plot(x, type = "line")
}
