% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spec_regarima.R
\name{set_transform}
\alias{set_transform}
\title{Set Transform Specification}
\usage{
set_transform(
  x,
  fun = c(NA, "Auto", "Log", "None"),
  adjust = c(NA, "None", "LeapYear", "LengthOfPeriod"),
  aicdiff = NA,
  fct = NA
)
}
\arguments{
\item{x}{the specification.}

\item{fun}{the transformation of the input series: \code{"None"} = no transformation of the series;
\code{"Log"} = takes the log of the series; \code{"Auto"} = the program tests for the log-level specification.}

\item{adjust}{(REGARIMA/X-13 specific) pre-adjustment of the input series for the length of period or leap year effects:
\code{"None"} = no adjustment; \code{"LeapYear"} = leap year effect; \code{"LengthOfPeriod"} = length of period.
Modifications of this variable are taken into account only when \code{function = "Log"}.}

\item{aicdiff}{(REGARIMA/X-13 specific)  a numeric defining the difference in AICC needed to accept no transformation when the automatic
transformation selection is chosen (considered only when \code{transform.function} is set to \code{"Auto"}).}

\item{fct}{(TRAMO specific) \code{numeric} controlling the bias in the log/level pre-test:
\code{ transform.fct }> 1 favours levels, \code{transform.fct}< 1 favours logs.
Considered only when \code{transform.function} is set to \code{"Auto"}.}
}
\description{
Set Transform Specification
}
