% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calendars.R
\name{single_day}
\alias{single_day}
\title{Single Date}
\usage{
single_day(date, weight = 1)
}
\arguments{
\item{date}{the date of the holiday in the format `"YYYY-MM-DD"`.}

\item{weight}{weight associated to the holiday.}
}
\description{
Single Date
}
\examples{
single_day("1999-03-19")
}
